Housing Prices
Housing Price Prediction using the Random Forest algorithm. This model takes into account about 1,500 different houses and all the features related to the house and then attempts to predict the cost. This model is highly inaccurate but it is one of my first projects and I hope to continue improving throughout.