Create Automating Superconducting Asynchronous Ballistic Circuit Discovery
Detailed is the final code for my senior design project which builds upon a started software tool that iterates over randomly created circuit components and values at an attempt of finding interesting circuits. These circuits are created in sponsorship from Sandia National Laboratories with aims at discovering circuits that can have some sort of unique value to the R&D of SABCs. The circuits are fed into a simulation application that then displays whether there is a presence of a flux quanta or not.